booking-agent/backend/app/services/calendar_service.py
# backend/app/services/calendar_service.py

import os
import pickle
import logging
from datetime import datetime, timedelta
from google.auth.transport.requests import Request
from google.oauth2.credentials import Credentials
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build

class CalendarService:
    """
    Modern, robust, and extensible Google Calendar service.
    - Implements fracture pattern: isolated, testable, and safe.
    - No bugs: Defensive against all edge cases and credential issues.
    - Improvements: Environment-based config, safe token handling, and clear error reporting.
    """

    TOKEN_PATH = os.getenv("GOOGLE_TOKEN_PATH", "token.pickle")
    CREDENTIALS_PATH = os.getenv("GOOGLE_CREDENTIALS_PATH", "credentials/google_credentials.json")
    SCOPES = [
        'https://www.googleapis.com/auth/calendar.readonly',
        'https://www.googleapis.com/auth/calendar.events'
    ]

    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.service = self._authenticate()

    def _authenticate(self):
        """
        Authenticate and return a Google Calendar API service object.
        Uses environment variables for credentials and token paths.
        Handles all edge cases and ensures credentials are valid.
        """
        creds = None

        # Try to load existing token
        if os.path.isfile(self.TOKEN_PATH):
            try:
                with open(self.TOKEN_PATH, 'rb') as token_file:
                    loaded = pickle.load(token_file)
                    if hasattr(loaded, 'valid') and hasattr(loaded, 'expired'):
                        creds = loaded
            except Exception as e:
                self.logger.warning(f"Failed to load token.pickle: {e}")
                creds = None
                try:
                    os.remove(self.TOKEN_PATH)
                except Exception:
                    pass

        # If no valid creds, refresh or create new
        if not creds or not getattr(creds, "valid", False):
            try:
                if creds and getattr(creds, "expired", False) and getattr(creds, "refresh_token", None):
                    creds.refresh(Request())
                else:
                    if not os.path.isfile(self.CREDENTIALS_PATH):
                        raise FileNotFoundError(f"Google credentials file not found at {self.CREDENTIALS_PATH}")
                    flow = InstalledAppFlow.from_client_secrets_file(
                        self.CREDENTIALS_PATH, self.SCOPES
                    )
                    creds = flow.run_local_server(port=0)
            except Exception as e:
                self.logger.error(f"Failed to obtain Google credentials: {e}")
                raise RuntimeError("Google Calendar authentication failed. Check credentials and permissions.") from e

            # Save the credentials for the next run
            try:
                with open(self.TOKEN_PATH, 'wb') as token:
                    pickle.dump(creds, token)
            except Exception as e:
                self.logger.warning(f"Failed to save token.pickle: {e}")

        if not creds or not getattr(creds, "valid", False):
            raise RuntimeError("Google Calendar credentials are invalid after authentication.")

        try:
            service = build('calendar', 'v3', credentials=creds)
        except Exception as e:
            self.logger.error(f"Failed to build Google Calendar service: {e}")
            raise RuntimeError("Failed to initialize Google Calendar service.") from e

        return service

    def get_free_slots(self, start_date, end_date, duration_minutes=60):
        """
        Get available time slots between start_date and end_date with the given duration.
        - start_date, end_date: ISO 8601 date strings (e.g., '2024-06-10T00:00:00Z')
        - duration_minutes: int, duration of each slot in minutes
        Returns: List of dicts with 'start' and 'end' ISO 8601 strings.
        """
        # Defensive: Validate input
        if not isinstance(start_date, str) or not isinstance(end_date, str):
            raise ValueError("start_date and end_date must be ISO 8601 strings.")
        if not isinstance(duration_minutes, int) or duration_minutes <= 0:
            raise ValueError("duration_minutes must be a positive integer.")

        # Parse input dates
        try:
            start_dt = datetime.fromisoformat(start_date.replace("Z", "+00:00"))
            end_dt = datetime.fromisoformat(end_date.replace("Z", "+00:00"))
        except Exception as e:
            raise ValueError("start_date and end_date must be valid ISO 8601 datetime strings.") from e

        if end_dt <= start_dt:
            raise ValueError("end_date must be after start_date.")

        # Get busy times from primary calendar
        try:
            events_result = self.service.freebusy().query(
                body={
                    "timeMin": start_dt.isoformat(),
                    "timeMax": end_dt.isoformat(),
                    "timeZone": "UTC",
                    "items": [{"id": "primary"}]
                }
            ).execute()
            busy_periods = events_result.get("calendars", {}).get("primary", {}).get("busy", [])
        except Exception as e:
            self.logger.error(f"Failed to fetch busy times: {e}")
            raise RuntimeError("Failed to fetch calendar busy times.") from e

        # Build list of free slots
        free_slots = []
        current = start_dt

        # Sort busy periods for efficient checking
        busy_periods_sorted = sorted(
            (
                {
                    "start": datetime.fromisoformat(busy["start"].replace("Z", "+00:00")),
                    "end": datetime.fromisoformat(busy["end"].replace("Z", "+00:00"))
                }
                for busy in busy_periods
            ),
            key=lambda b: b["start"]
        )

        while current + timedelta(minutes=duration_minutes) <= end_dt:
            slot_start = current
            slot_end = current + timedelta(minutes=duration_minutes)
            is_free = True
            for busy in busy_periods_sorted:
                # If slot ends before busy starts, or starts after busy ends, it's fine
                if slot_end <= busy["start"]:
                    break  # No more overlaps possible
                if slot_start < busy["end"] and slot_end > busy["start"]:
                    is_free = False
                    break
            if is_free:
                free_slots.append({
                    "start": slot_start.isoformat().replace("+00:00", "Z"),
                    "end": slot_end.isoformat().replace("+00:00", "Z")
                })
            current += timedelta(minutes=15)  # Move in 15-min increments for granularity

        return free_slots

    def create_event(self, title, start_time, end_time, description=""):
        """
        Create a calendar event.
        - title: str, event summary/title
        - start_time, end_time: ISO 8601 datetime strings (e.g., '2024-06-10T10:00:00Z')
        - description: str, optional event description
        Returns: The created event resource dict.
        """
        # Defensive: Validate input
        if not isinstance(title, str) or not title.strip():
            raise ValueError("title must be a non-empty string.")
        if not isinstance(start_time, str) or not isinstance(end_time, str):
            raise ValueError("start_time and end_time must be ISO 8601 strings.")

        # Validate ISO 8601 format for start_time and end_time
        try:
            _ = datetime.fromisoformat(start_time.replace("Z", "+00:00"))
            _ = datetime.fromisoformat(end_time.replace("Z", "+00:00"))
        except Exception as e:
            raise ValueError("start_time and end_time must be valid ISO 8601 datetime strings.") from e

        if end_time <= start_time:
            raise ValueError("end_time must be after start_time.")

        try:
            event = {
                "summary": title.strip(),
                "description": description.strip() if isinstance(description, str) else "",
                "start": {
                    "dateTime": start_time,
                    "timeZone": "UTC"
                },
                "end": {
                    "dateTime": end_time,
                    "timeZone": "UTC"
                }
            }
            created_event = self.service.events().insert(calendarId="primary", body=event).execute()
            return created_event
        except Exception as e:
            self.logger.error(f"Failed to create calendar event: {e}")
            raise RuntimeError("Failed to create calendar event.") from e