booking-agent/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.3.0
anthropic>=0.8.0
sqlalchemy>=2.0.23
alembic>=1.12.1
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.2
pytz>=2023.3
google-auth>=2.40.0